requirements.txt
pyserial>=3.5
requests>=2.31.0
EAS2Text-Remastered>=1.0.0
